Database update issues is resolved
The calls to LiveData where not producing and value or most null,
It seemed like the data was not being populated into the database
at all, which was the case.
Am not clear on what the issues was, my guesses point to
be an instance problem or thread issue was the cause.